pyproject.toml
[project]
name = "langchain-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
